+++ src/components/NewTransducer/NewTransducer.module.css
@@ -1,0 +1,80 @@
+.form_container {
+  margin: 2rem;
+}
+
+.header {
+  color: #a6518ff1;
+  font-weight: bolder;
+  margin-bottom: 1rem;
+}
+
+.form_fieldset {
+  padding: 0.60rem;
+}
+
+.legend {
+  padding: 0 0.30rem;
+  color: #a6518ff1;
+  font-weight: bold;
+}
+
+.field {
+  margin: 1rem;
+  margin-top: 0.75rem;
+  font-size: large;
+}
+
+.field label {
+  display: block;
+  color: #ffebcd;
+  font-weight: bolder;
+}
+
+.field input,
+.field textarea,
+.field select {
+  width: 100%;
+  padding: 0.30rem;
+  border: none;
+  border-bottom: 2px solid pink;
+  background-color: #773f6828;
+  color: #fff;
+  font-size: large;
+}
+
+.field textarea {
+  resize: none;
+  font-size: large;
+}
+
+.field select {
+  cursor: pointer;
+}
+
+.checkbox_field {
+  display: flex;
+  justify-content: center;
+}
+
+.checkbox_field label {
+  color: pink;
+  font-weight: bolder;
+}
+
+.checkbox_field input {
+  margin-right: 0.50rem;
+}
+
+.line_break {
+  margin: 1.50rem 0;
+}
+
+.form_actions {
+  display: flex;
+  justify-content: right;
+  margin-top: 1rem;
+}
+
+.form_actions button {
+  margin-left: 0.75rem;
+}
