+++ src/components/Transducers/Transducers.tsx
@@ -1,30 +1,30 @@
 import { useState, useRef, useContext } from 'react';
 
+import { TransducerContext, TransducerContextType } from '../../store/transducer-context';
 import Modal, { ModalHandle } from '../../ui/Modal/Modal';
 import TransducerItem from '../TransducerItem/TransducerItem';
 import { Transducer } from '../../data/data';
 import FullDetails from '../FullDetails/FullDetails';
-import { TransducerContext, TransducerContextType } from '../../store/transducer-context';
 import styles from './Transducers.module.css';
 
 const Transducers = () => {
   const { transducers } = useContext<TransducerContextType>(TransducerContext);
   const [selectedTransducer, setSelectedTransducer] = useState<Transducer | undefined>();
-  const dialog = useRef<ModalHandle>();
+  const modalRef = useRef<ModalHandle>();
 
   const handleClickedTransducer = (selectedTransducer: Transducer) => {
     setSelectedTransducer(selectedTransducer);
-    dialog.current.open();
+    modalRef.current.open();
   };
 
-  const handleCloseModal = () => {
-    dialog.current.close();
+  const handleCloseDetails = () => {
+    modalRef.current.close();
   };
 
   return (
     <>
-      <Modal ref={dialog}>
-        {selectedTransducer && <FullDetails transducer={selectedTransducer} onCloseModal={handleCloseModal} />}
+      <Modal ref={modalRef}>
+        {selectedTransducer && <FullDetails transducer={selectedTransducer} onCloseModal={handleCloseDetails} />}
       </Modal>
       <ul className={styles.container}>
         {transducers.map((transducer: Transducer) => (
